- обновление формы процесса,  - обновление тестовых данных

+++ web-client/src/main/java/ru/codeinside/gses/activiti/ftarchive/style/TableStyle.java
@@ -2,7 +2,7 @@ package ru.codeinside.gses.activiti.ftarchive.style;
 
 public class TableStyle {
 
-  public static float DATA_COL_WIDTH = 600;
-  public static float BUTTON_COL_WIDTH = 100;
+  public static int DATA_COL_WIDTH = 600;
+  public static int BUTTON_COL_WIDTH = 100;
 
 }
